HSI-1000 [BE] No se podrán borrar los documentos de consentimiento

+++ back-end/hospital-api/src/main/java/net/pladema/staff/controller/mapper/BackofficeEpisodeDocumentTypeStore.java
@@ -66,7 +66,12 @@ public class BackofficeEpisodeDocumentTypeStore implements BackofficeStore<Episo
 
 	@Override
 	public void deleteById(Integer id) {
-		repository.deleteById(id);
+		repository.findById(id).ifPresent(episodeDocumentType -> {
+			if (episodeDocumentType.getConsentId().equals(1))
+				repository.deleteById(id);
+			else
+				throw new BackofficeValidationException("El documento es de consentimiento y no se puede eliminar");
+		});
 	}
 
 	@Override
